+++ system/modules/newsletter/modules/ModuleNewsletterList.php
@@ -98,7 +98,7 @@ class ModuleNewsletterList extends \Module
 
 						if ($objJumpTo !== null)
 						{
-							$arrJumpTo[$objTarget->jumpTo] = $this->generateFrontendUrl($objJumpTo->row(), ($GLOBALS['TL_CONFIG']['useAutoItem'] ?  '/%s' : '/items/%s'));
+							$arrJumpTo[$objTarget->jumpTo] = $this->generateFrontendUrl($objJumpTo->row(), (($GLOBALS['TL_CONFIG']['useAutoItem'] && !$GLOBALS['TL_CONFIG']['disableAlias']) ?  '/%s' : '/items/%s'));
 						}
 						else
 						{